Size GUIstyle surfaces and textures based on the actual scale. Give Font the responsibility for rendering text FontManager is only responsible for managing the font cache.

+++ include/misc/string_util.h
@@ -186,7 +186,7 @@ std::string utf8Substr(std::string_view str, size_t pos, size_t len = std::strin
    into multiple lines.
 */
 std::vector<std::string>
-greedyWordWrap(std::string_view text, int linewidth, std::function<int(std::string_view)> pGetTextWidth);
+greedyWordWrap(std::string_view text, float linewidth, std::function<float(std::string_view)> pGetTextWidth);
 
 std::string convertCP850ToUTF8(std::string_view text);
